refactor: merge lotto generator into lotto machine
